Refactor NVD initialization, set it up for hot reloading with updates (#62)
Get rid of the vulnmdsrc abstraction. NVD is the only real metadata source, and I didn't want us to download data from them multiple times just to avoid breaking the abstraction.
Refactor Clair's NVD updater to take a directory of JSON files. Move the download of the JSON files to a separate helper in pkg. This allows us to download the JSON files once and use them for both Clair's updater and to update our NVD definitions.
Move the singleton cache and cveMap into a separate package and protect them with a lock. This will allow us to hot-replace them with new vuln definitions at will.