Make the RHEL updater much faster (#82)
Got annoyed enough at how RHEL kept throttling us, so fixed the updater to fetch the all.xml file. However, there are some vulns which are not included in this file, and I didn't want to lose any functionality, so continue to fetch vulns not included in this manually. Now we make 300 requests instead of 3400. I verified with two JSON dumps (a before and an after) that the results were exactly identical (modulo sorting issues).