Run the diff command in CI (#81)
Add a genesis_manifests.json file which contains all known genesis dumps, and the locations of corresponding diff dumps. (This is used when uploading the files, but will also be baked into the scanner image for the scanner's uploader.)
In CI, pull the latest dump updates, generate diffs with all known genesis files, and upload them to the definitions.stackrox.io bucket.
Also take a postgres dump and NVD dump from the latest updates, and upload them to the scanner CI bucket. This will be used directly in PR builds, to avoid pulling dumps all the time. This is already being used in scanner builds after #84 was merged.