Fix downstream dumps upload on tagged build
Blobs upload job did not find blobs in the workspace
https://app.circleci.com/pipelines/github/stackrox/scanner/19362/workflows/a0dca865-c4a7-409b-a168-582659cdca0d/jobs/199516?invite=true#step-105-43

because the step to prepare them was skipped
https://app.circleci.com/pipelines/github/stackrox/scanner/19362/workflows/a0dca865-c4a7-409b-a168-582659cdca0d/jobs/199503?invite=true#step-102-25

This happened due to `upload-dumps-for-embedding-into-image` being
skipped on tagged builds.
This job is already enabled for master branch and does upload only
for master branch. Therefore enabling tagged builds should
effectively allow it to run for `master` branch when the commit is
tagged.