Add RHTAP build pipeline
A few differences from the standard RHTAP build pipeline:

* Use of 6GB buildah image to avoid OOMs during container build task
* There is an extra step to fetch the vuln feed data.  This is done
  outside of the build step to make hermetic builds easier.

The vuln feed data script pulls the definitions from the Google storage
location and writes them to the "source" folder of the "source"
workspace which is shared by the build-container task.  This task will
build using `image/scanner/rhtap/Dockerfile`, which expects the vuln
files to be in the buildah working directory.

I decided to create a separate dockerfile for RHTAP to avoid any
regressions with modifying the existing ones.  The dockerfile was
created using a combination of the upstream and downstream dockerfiles.

Current RHTAP build trigger config:

* PRs will only build in RHTAP when "rhtap" is in the branch name
* Pushes to master will trigger an RHTAP build