Switch base stage image tag to `latest`
Clair scan finds CVEs in the currently used image which has a pinned tag `registry.redhat.io/rhel8/postgresql-12:1-150.1698060555`.

Previously, I insisted and we decided :-) that it's ok to unpin tags on certain Red Hat images and use `latest` instead.

First, Red Hat base images such as `ubi8`/`rhel8` are quite stable and we did not experience anything breaking when switching to newer tags yet. I hope this applies to `rhel8/postgresql-*` just as well.

Second, there is (was) no mechanism to auto-update tags in Konflux, and so the only way to get auto-update effect is by using floating tags.