Rename Dockerfile.konflux -> konflux.Dockerfile
GoLand considers (at least at some point did) `Dockerfile` to be an
extension and so Dockerfile syntax highlighting works only when the
file has `.Dockerfile` at the end.
Obviously, `konflux.Dockerfile.slim` won't get this but I don't want
to break the pattern with the existing `Dockerfile.slim`.

Hopefully, this is not too confusing.