ROX-26784: Only trigger Konflux builds against master
This is primarily intended to stop triggering builds in Konflux for "master" components when there are PRs or pushes on release branches.

This retains the branch naming and label conventions to reduce load on Konflux.

One side effect of this change is that PRs against branches other than master or release branches will not have Konflux builds ever run against them.